% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/frenum_miss.cpp
\name{frenum_miss}
\alias{frenum_miss}
\title{Fast renumbering of missing id}
\usage{
frenum_miss(id)
}
\arguments{
\item{id}{a numeric or character vector for which the missing codes will be
replaced with NA}
}
\value{
a numeric or character vector
}
\description{
\code{frenum_miss()} will replace certain values with \code{NA}. For numeric vectors
ids coded as \code{0} or \code{-99} will be considered as missing or unknown. For
character vectors this is extended to include \code{'.'}, \code{'*'}, \code{'NA'} and \code{' '}.
}
\examples{
set.seed(42L)
ids <- rpois(50, 100)
ids[sample.int(50, 10)] <- c(0L, -99)[sample.int(2, 10, replace = TRUE)]
frenum_miss(ids)

ids <- c("A", "B", ".", "D", "*", "NA", " ", "  ", "K")
frenum_miss(ids)

}
\seealso{
\code{\link[=renum_miss]{renum_miss()}} which is the R implementation of the same
}
