% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{msg_wrap}
\alias{msg_wrap}
\title{Wrap messages for warnings, error or diagnostic messages}
\usage{
msg_wrap(...)
}
\description{
\code{msg_wrap()} is a utility function that can be used when messages intended
for \code{\link[=warning]{warning()}}, \code{\link[=stop]{stop()}} or \code{\link[=message]{message()}} get too long. Informative messages
are useful for the users but clog the script for a developer.
}
\keyword{internal}
