% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_add.R
\name{ped_add}
\alias{ped_add}
\title{Add parents to pedigree}
\usage{
ped_add(pedigree, ...)
}
\arguments{
\item{pedigree}{an object coercible to a \code{data.table} which includes the
id of individual, sire and dam.}
}
\value{
a \code{data.table} containing the missing parents
}
\description{
\code{ped_add()} takes an \code{R} object and assumes the first three elements
corresponds to the id of the individual, sire and dam, in that order. If the
\code{pedigree} object contains additional columns these will be filled with \code{NA}
by default for the added individuals.
}
\examples{

}
