% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_prep.R
\name{ped_prep}
\alias{ped_prep}
\title{Preprocess pedigree}
\usage{
ped_prep(pedigree, na_id_rm = FALSE)
}
\arguments{
\item{pedigree}{an object coercible to a \code{data.table} which includes the
id of individual, sire and dam.}

\item{na_id_rm}{a logical indicating whether to remove individuals coded as
missing or unkown. When set to \code{FALSE} it will return an error.}
}
\value{
a \code{data.table} containing the original ids and new id columns used
in further calculation steps
}
\description{
\code{ped_prep()} prepares the given pedigree by recoding missing ids to \code{NA},
checking for duplicate records and some more checks. This is usually the
first step before any further calculations are performed.
}
\details{
The first three columns are assumed to be the id of the individual, sire and
dam, respectively in that order. New columns .id, .sire and .dam are created
which are used by other functions in this package.
}
\seealso{
\code{\link[=ped_add]{ped_add()}} is a function that relies on preprocessing before adding
the missing parents, for example.
}
