% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_list.R
\name{str_list}
\alias{str_list}
\title{List items}
\usage{
str_list(..., sep = ", ", last = " and ")
}
\arguments{
\item{...}{one or more character vectors or coercible to character vector.}

\item{sep}{a string to seperate elements except last.}

\item{last}{a string to seperate last element.}
}
\value{
a character string.
}
\description{
`str_list()` wraps [base::paste0()] in order to turn a list of character
strings into a grammatically correct list without an
[Oxford Comma](https://en.wikipedia.org/wiki/Serial_comma) by default.
Sue me.
}
\examples{
shopping_list <- c("bread", "milk", "eggs")
str_list(shopping_list)
}
