% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renum_missing.R
\name{renum_miss}
\alias{renum_miss}
\title{Renumber missing id}
\usage{
renum_miss(id, warn = TRUE)
}
\arguments{
\item{id}{a numeric or character vector for which the missing codes will be
replaced with NA}

\item{warn}{a logical indicating wheter to print a warning message with the
codes that will be considered as missing}
}
\value{
a numeric of character vector where the codes for missing ids have
  been replaced with NA
}
\description{
Renumber missing id
}
\examples{
set.seed(42L)
ids <- rpois(50, 100)
ids[sample.int(50, 10)] <- c(0L, -99)[sample.int(2, 10, replace = TRUE)]
renum_miss(ids)
}
